requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
motor==3.1.1
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==1.10.7
email-validator==2.0.0